refactor(sdk-trace): align benchmark scripts with the rest of the repo
