+++ packages/sdk-trace/test/performance/benchmark/index.js
@@ -3,4 +3,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-require('./sampler');
+require('./random-id-generator');
+require('./span');
+require('./batch-span-processor');
