+++ packages/sdk-trace/test/performance/benchmark/micro-benchmarks.js
@@ -1,0 +1,7 @@
+/*
+ * Copyright The OpenTelemetry Authors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+require('./sampler');
+require('./setAttributes');
